third_party/blink/web_tests/plugins/mouse-click-iframe-to-plugin-expected.txt

Plugin received event: MouseEnter
Plugin received event: MouseMove
Plugin received event: MouseDown
Plugin received event: MouseUp


PASSED
